+++ src/harlequin/__init__.py
@@ -1,1 +1,14 @@
-from harlequin.app import Harlequin as Harlequin
+from harlequin.adapter import HarlequinAdapter, HarlequinConnection, HarlequinCursor
+from harlequin.app import Harlequin
+from harlequin.autocomplete import HarlequinCompletion
+from harlequin.options import HarlequinAdapterOption, HarlequinCopyFormat
+
+__all__ = [
+    "Harlequin",
+    "HarlequinAdapter",
+    "HarlequinConnection",
+    "HarlequinCursor",
+    "HarlequinCopyFormat",
+    "HarlequinAdapterOption",
+    "HarlequinCompletion",
+]